removed EnumPair and LogLevel
EnumPair was only used at microchip to specify the firmware message
which was two numbers <module>.<message>, and you could specify a string
or a number for each part, but the second set of strings available depended
on which module you had typed.  More confusing and probably never used again.